requirements.txt
playwright
readability-lxml
beautifulsoup4
requests
transformers
torch
chromadb
gTTS
python-dotenv
tqdm
colorama>=0.4.6
requests>=2.31.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
tqdm>=4.66.0
colorama>=0.4.6
accelerate>=0.27.2